Add services preferences

+++ app/src/main/java/io/github/wulkanowy/ui/modules/main/MainActivity.kt
@@ -125,7 +125,7 @@ class MainActivity : BaseActivity(), MainView {
         presenter.onBackPressed { super.onBackPressed() }
     }
 
-    override fun startSyncService(interval: Int, useOnlyWifi: Boolean) {
+    override fun startSyncService(interval: Long, useOnlyWifi: Boolean) {
         JobManager().start(interval, useOnlyWifi)
     }
